revisi again..
